// app/types/property.ts
export interface Property {
  id: number;
  address: string;
  name: string;
  company_name: string;
  phone_number: string;
  locality: string;
  url: string;
  image_url: string;
  list_price: string;
  type: string;
  bedrooms: number;
  bathrooms: number;
  sq_ft: number;
  date: string;
  source: string;
}